requirements.txt
yfinance>=0.2.0
langchain-groq>=0.2.0
python-dotenv>=1.0.0
pandas>=2.0.0
